+++ lib/date/index.test.js
@@ -4,12 +4,11 @@
  */
 'use strict';
 
-var dateFormatter = require('./index.js');
-var MyDate = require('./index.js');
+var dateFormatter = require('./dateFormatter.js');
+var dateCompare = require('./dateCompare.js');
 var expect = require('chai').expect;
 
-// console.log('MyDate:',MyDate);
-describe(' format 测试', function() {
+describe(' dateFormatter 测试', function() {
   var case_data = {
     'timeSecondsFormatter': {
       param: {
@@ -59,100 +58,50 @@ describe(' format 测试', function() {
         format: 'yyyy-M-dTh:m:s'
       },
       result: '2016-7-4T18:22:21'
-    },
-    'stringFormmaterComplete': {
-      param: {
-        date: '16-01-01',
-        format: 'yyyy-MM-dd'
-      },
-      result: '2016-01-01'
-    },
+    }
   }
 
 
   for (var name in case_data) {
     it(name + ' 返回' + JSON.stringify(case_data[name].result), function() {
-      var myDate = new MyDate(case_data[name].param.date);
-      var result = myDate.format(case_data[name].param.format)
-      expect(result).to.be.deep.equal(case_data[name].result);
+      expect(dateFormatter(case_data[name].param.date, case_data[name].param.format)).to.be.deep.equal(case_data[name].result);
     });
   }
 
 });
 
 
-describe(' getTimeUTC 测试', function() {
+describe(' dateCompare 测试', function() {
   var case_data = {
-    'getTimeUTC ': {
-      param: {
-        date: 1467627741105
-      },
-      result: 1467598941105
-    }
-  }
-
-  for (var name in case_data) {
-    it(name + ' 返回' + JSON.stringify(case_data[name].result), function() {
-      var myDate = new MyDate(case_data[name].param.date);
-      var result = myDate.getTimeUTC();
-      // console.log('98:',MyDate.getzoneOffset());
-      expect(result).to.be.deep.equal(case_data[name].result);
-    });
-  }
-})
-
-describe(' changeTimezone 测试', function() {
-  var case_data = {
-    'changeTimezone  ': {
-      param: {
-        date: 1467725512621,
-        zon1: 8,
-        zon2: 0
-      },
-      result: 1467696712621
-    }
-  }
-
-  for (var name in case_data) {
-    it(name + ' 返回' + JSON.stringify(case_data[name].result), function() {
-      var myDate = new MyDate(case_data[name].param.date);
-      myDate.changeTimezone(case_data[name].param.zon1, case_data[name].param.zon2);
-      expect(myDate.getTime()).to.be.deep.equal(case_data[name].result);
-    });
-  }
-})
-
-describe(' derTime 测试', function() {
-  var case_data = {
-    'getTimeUTC hour ': {
+    'dateCompare hour ': {
       param: {
         date1: 1467627741105,
         date2: 1467598941105
       },
       result: '8小时后'
     },
-    'getTimeUTC just ': {
+    'dateCompare just ': {
       param: {
         date1: 1467627741105,
         date2: 1467627741105
       },
       result: '刚刚'
     },
-    'getTimeUTC minute ': {
+    'dateCompare minute ': {
       param: {
         date1: 1467627000000,
         date2: 1467627741105
       },
       result: '12分钟前'
     },
-    'getTimeUTC month ': {
+    'dateCompare month ': {
       param: {
         date1: 1467627000000,
         date2: 1464948600000
       },
       result: '1个月后'
     },
-    'getTimeUTC year ': {
+    'dateCompare year ': {
       param: {
         date1: 1467627000000,
         date2: 1436091000000
@@ -163,9 +112,7 @@ describe(' derTime 测试', function() {
 
   for (var name in case_data) {
     it(name + ' 返回' + JSON.stringify(case_data[name].result), function() {
-      var myDate = new MyDate(case_data[name].param.date1);
-      var result = myDate.derTime(case_data[name].param.date1, case_data[name].param.date2);
-      expect(result).to.be.deep.equal(case_data[name].result);
+      expect(dateCompare(case_data[name].param.date1, case_data[name].param.date2)).to.be.deep.equal(case_data[name].result);
     });
   }
 })
