--- lib/myDate/README.md
@@ -1,8 +1,0 @@
-## DateUtil
-
-> 将Date数据转换成想要的格式
-
-```javascript
-var dateString = dateFormatter(1467627741105,'yyyy-MM-dd')
-console.log(dateString); 
-```
