--- lib/sort/README.md
@@ -1,6 +1,0 @@
-## ParseUrl
-
-> 以各种排序方法排序
-
-```javascript
-```
